requirements.txt
holoviews
geoviews
panel
blackboxprotobuf
pycryptodome
jupyter-bokeh
geopandas==1.0.0
fiona
typer
